Refactor of DatHDF and get_dat now loads safer
get_dat (from_exp_filepath) now uses a global lock for the file when copying from exp datapath to save_path
